refactor: extract service bootstrap from server/index.ts (#442)
Move all service initialization, cross-dependency wiring, and shutdown
registration into server/bootstrap.ts. This reduces the god module from
987 to 736 lines (-25%) and makes the dependency graph explicit via the
ServiceContainer interface. Part 1 of the #442 decomposition.
